Redone structure for modules
