+++ Modules/date.js
@@ -1,4 +1,4 @@
-import { DateTime } from "../node_modules/luxon/src/luxon.js";
+import { DateTime } from '../node_modules/luxon/src/luxon.js';
 
 const current = DateTime.now().toString();
 document.getElementById('date').innerHTML = current;
