+++ index.html
@@ -13,9 +13,9 @@
 			<div class="nav-name">
 				<p>Awesome Books</p>
 			</div>
-			<a href="#" class="nav-link" id="nav-list">List</a>
-			<a href="#" class="nav-link" id="nav-add">Add new</a>
-			<a href="#" class="nav-link" id="nav-contact">Contact</a>
+			<a href="#" class="nav-link" onclick = "showList()" id="nav-list">List</a>
+			<a href="#" class="nav-link" onclick = "showAdd()" id="nav-add">Add new</a>
+			<a href="#" class="nav-link" onclick = "showContactPage()" id="nav-contact">Contact</a>
 		</div>
 	</header>
 	<div class="date-container">
@@ -72,6 +72,6 @@
 	<footer>
 		<h2>Copyright...</h2>
 	</footer>
-	<script src="index.js"></script>
+	<script src="index.js" type="module"></script>
 </body>
 </html>
